Split web main module
