+++ apps/web/src/sorters.js
@@ -1,0 +1,7 @@
+export function sortNewest(a, b) {
+  return b.createdAt.localeCompare(a.createdAt);
+}
+
+export function sortWorkers(a, b) {
+  return `${a.gpuId}-${a.id}`.localeCompare(`${b.gpuId}-${b.id}`);
+}
